requirements.txt
Adafruit-Blinka==8.18.1
adafruit-circuitpython-busdevice==5.2.4
adafruit-circuitpython-lps2x==3.0.1
adafruit-circuitpython-neopixel==6.3.8
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-register==1.9.15
adafruit-circuitpython-requests==1.13.2
adafruit-circuitpython-typing==1.9.1
Adafruit-PlatformDetect==3.45.1
Adafruit-PureIO==1.1.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
gTTS==2.3.1
idna==3.4
numpy==1.24.3
pandas==2.0.0
Pillow==9.5.0
PyAudio==0.2.13
pyftdi==0.54.0
pygame==2.3.0
pyserial==3.5
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2023.3
pyusb==1.2.1
requests==2.28.2
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
six==1.16.0
SpeechRecognition==3.10.0
sysv-ipc==1.1.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
